+++ grpc-protocol/src/main/java/com/linecorp/armeria/common/grpc/protocol/AbstractUnaryGrpcService.java
@@ -18,20 +18,7 @@ package com.linecorp.armeria.common.grpc.protocol;
 
 import java.util.concurrent.CompletableFuture;
 
-import com.linecorp.armeria.common.HttpData;
-import com.linecorp.armeria.common.HttpHeaderNames;
-import com.linecorp.armeria.common.HttpHeadersBuilder;
-import com.linecorp.armeria.common.HttpRequest;
-import com.linecorp.armeria.common.HttpResponse;
-import com.linecorp.armeria.common.HttpStatus;
-import com.linecorp.armeria.common.ResponseHeaders;
-import com.linecorp.armeria.common.grpc.protocol.ArmeriaMessageDeframer.ByteBufOrStream;
-import com.linecorp.armeria.common.grpc.protocol.ArmeriaMessageDeframer.Listener;
-import com.linecorp.armeria.server.AbstractHttpService;
-import com.linecorp.armeria.server.ServiceRequestContext;
-
 import io.netty.buffer.ByteBuf;
-import io.netty.buffer.ByteBufAllocator;
 import io.netty.buffer.ByteBufUtil;
 import io.netty.buffer.Unpooled;
 
@@ -42,12 +29,7 @@ import io.netty.buffer.Unpooled;
  * <p>This service does not support compression. If you need support for compression, please consider using
  * normal gRPC stubs or file a feature request.
  */
-public abstract class AbstractUnaryGrpcService extends AbstractHttpService {
-
-    private static final ResponseHeaders RESPONSE_HEADERS =
-            ResponseHeaders.of(HttpStatus.OK,
-                               HttpHeaderNames.CONTENT_TYPE, "application/grpc+proto",
-                               GrpcHeaderNames.GRPC_ENCODING, "identity");
+public abstract class AbstractUnaryGrpcService extends AbstractUnsafeUnaryGrpcService {
 
     /**
      * Returns an unframed response message to return to the client, given an unframed request message. It is
@@ -57,67 +39,13 @@ public abstract class AbstractUnaryGrpcService extends AbstractHttpService {
     protected abstract CompletableFuture<byte[]> handleMessage(byte[] message);
 
     @Override
-    protected final HttpResponse doPost(ServiceRequestContext ctx, HttpRequest req) {
-        final CompletableFuture<HttpResponse> responseFuture =
-                req.aggregateWithPooledObjects(ctx.contextAwareEventLoop(), ctx.alloc())
-                   .thenCompose(msg -> deframeMessage(msg.content(), ctx.alloc()))
-                   .thenCompose(requestMessage -> {
-                       final byte[] requestMessageBytes;
-                       try {
-                           requestMessageBytes = ByteBufUtil.getBytes(requestMessage);
-                       } finally {
-                           requestMessage.release();
-                       }
-                       return handleMessage(requestMessageBytes);
-                   })
-                   .thenApply(responseMessage -> {
-                       final ArmeriaMessageFramer framer = new ArmeriaMessageFramer(
-                               ctx.alloc(), Integer.MAX_VALUE);
-                       final HttpData framed = framer.writePayload(Unpooled.wrappedBuffer(responseMessage));
-                       return HttpResponse.of(
-                               RESPONSE_HEADERS,
-                               framed,
-                               GrpcTrailersUtil.statusToTrailers(StatusCodes.OK, null, true).build());
-                   })
-                   .exceptionally(t -> {
-                       final HttpHeadersBuilder trailers;
-                       if (t instanceof ArmeriaStatusException) {
-                           ArmeriaStatusException statusException = (ArmeriaStatusException) t;
-                           trailers = GrpcTrailersUtil.statusToTrailers(
-                                   statusException.getCode(), statusException.getMessage(), false);
-                       } else {
-                           trailers = GrpcTrailersUtil.statusToTrailers(
-                                   StatusCodes.INTERNAL, t.getMessage(), false);
-                       }
-                       return HttpResponse.of(trailers.build());
-                   });
-
-        return HttpResponse.from(responseFuture);
-    }
-
-    private CompletableFuture<ByteBuf> deframeMessage(HttpData framed, ByteBufAllocator alloc) {
-        final CompletableFuture<ByteBuf> deframed = new CompletableFuture<>();
-        try (ArmeriaMessageDeframer deframer = new ArmeriaMessageDeframer(
-                new Listener() {
-                    @Override
-                    public void messageRead(ByteBufOrStream message) {
-                        // Compression not supported.
-                        assert message.buf() != null;
-                        deframed.complete(message.buf());
-                    }
-
-                    @Override
-                    public void endOfStream() {
-                        if (!deframed.isDone()) {
-                            deframed.complete(Unpooled.EMPTY_BUFFER);
-                        }
-                    }
-                },
-                Integer.MAX_VALUE,
-                alloc)) {
-            deframer.request(1);
-            deframer.deframe(framed, true);
+    protected final CompletableFuture<ByteBuf> handleMessage(ByteBuf message) {
+        final byte[] bytes;
+        try {
+            bytes = ByteBufUtil.getBytes(message);
+        } finally {
+            message.release();
         }
-        return deframed;
+        return handleMessage(bytes).thenApply(Unpooled::wrappedBuffer);
     }
 }
